refactor(submit): extract handlers into submit_handlers.go
Move the submit Handler implementations out of submit.go into
submit_handlers.go, matching the convention every other command follows
(sync_handlers.go, scope_handlers.go, ...) and the docs/tui.md handler
pattern. submit.go now holds only the command wiring (flags, executeSubmit,
NewSubmitCmd, NewSsCmd); NewSubmitUI, both handlers, and snapshotToStackTree
live alongside the existing submit_handlers_test.go.

SimpleSubmitHandler now embeds cli/common.BaseHandler instead of rolling
its own output/mutex, dropping the dead out (io.Writer) and started fields
left over from the compact refactor.

Also make "create" a named ActionCreate constant in the submit component
to satisfy goconst.
